Fix class name check when this project is built
When this package is built, the value of `this.constructor.name` is minified (something like "j0") and thus the `indexOf` check always returns false. This change moves it to a property on the class so that the build can't mess with it.

This commit should fix this downstream issue: https://github.com/raineorshine/npm-check-updates/issues/1460